requirements.txt
flask>=3.0
python-dotenv>=1.0
numpy>=2.0
pillow>=10.0
scipy>=1.13
